Cache access to arrays to improve performance
It turns out, when accessing arrays repeatedly, Hotspot doesn't cache
the value fetched from the array. This is likely because Hotspot cannot
assume the array won't be modified somewhere else in between calls.

I don't know of an annotation similar to `restrict` in C which would
indicate to Hotspot that it's safe to cache the value fetched from the
array. Until then, do it manually.