Ensure every iterations of each benchmark runs on the same data
Because the BLAS operations update the data in place, each iteration of
each benchmark would end up operating on a different set of data. It was
particularely problematic for operations like dscal where all values
would converge to 0, leading to skewed results. Instead, we clone the
data that's going to be overwritten, and pass a fresh copy every time.