Split different cases in dedicated methods
Larger operations like dgemm and dsymm would reach such a large size
that it wouldn't get compiled by Hotspot. The simplest solution is to
split up the methods into the different cases it deals with. This also
has the advantage that it's easier to override in inheriting classes
like VectorizedBLAS.