Reintroduce simple dgemm and sgemm version for small matrices
The GOTO implementation would be much slower for small matrices, where
the cost of copying isn't amortized by the larger number of operations.
From local experiment, the threshold at which the GOTO implementation
gets faster is when the product of m, n, and k reaches 1e6, or matrices
of around 100 by 100 elements.