Add skeleton of support for ARPACK and LAPACK
This still doesn't compile. First, I need to have a proper interface to
express the API of LAPACK and ARPACK. Then, I'll want to implenent the
respective NetlibWrapper classes, and finally the JavaLAPACK and
JavaARPACK.

I'm not planning to add any test or benchmarks right away, as I do not
expect to have custom implementations of any of the operations just yet.